test(daemon): make the nightly test binary the nightly category, drop the nextest allowlist
nextest has no test tags; its categorisation units are package, binary and
test name. The `nightly` profile in .config/nextest.toml was a hand-maintained
list of test names, needed only because `#[ignore]` meant both "slow" and
"known-failing" in the same binary.

Move the five slow-but-passing tiers into tests/nightly.rs, with the shared
bodies in tests/common/scale.rs, so the binary itself is the category. The
known-failing H2 and M2 tests stay `#[ignore]`d next to their siblings, so they
can never be swept into the nightly run. The workflow now runs
`cargo nextest run -p atuin-daemon --test nightly --run-ignored all` and the
config file is gone.
