--- .config/nextest.toml
@@ -1,29 +1,0 @@
-# cargo-nextest configuration.
-#
-# The `nightly` profile below IS the "nightly category": the long-running tests that pass but are
-# too slow for per-push CI (1M rows / 100k record writes take minutes in a debug build). They are
-# `#[ignore]`d in the source, and the scheduled `.github/workflows/nightly.yml` job runs them with:
-#
-#     cargo nextest run -p atuin-daemon --profile nightly --run-ignored ignored-only
-#
-# To add a long test to the nightly run: `#[ignore]` it in the source, then add one
-# `test(=<its_name>)` line to the allowlist below.
-#
-# The allowlist is intentional and explicit: the other `#[ignore]`d tests document *unfixed*
-# defects and are expected to fail, so they must stay OUT of the nightly run --
-#   - `deleting_everything_in_one_request_{two_hundred_k,one_m}`: H2, a whole-batch delete past
-#     tonic's 4 MiB decode limit is rejected before any row is deleted (needs client-side chunking).
-#   - the `*_reload_*` / `concurrent_deletes_both_leave_the_index` concurrency tests: the M2
-#     index-reload lost-update.
-# When one of those defects is fixed, move its tier into this allowlist.
-
-[profile.nightly]
-default-filter = """
-    binary(scale) & (
-        test(=index_matches_db_after_seeding_one_m)
-        | test(=deleting_a_sample_leaves_everything_else_one_m)
-        | test(=deleting_everything_in_one_request_hundred_k)
-        | test(=rebuild_from_a_large_store_hundred_k)
-        | test(=rebuild_from_a_large_store_one_m)
-    )
-"""
